daoInterfaces

+++ Jakyjty/src/main/java/com/jakyjty/dao/IUserDao.java
@@ -1,9 +1,10 @@
 package com.jakyjty.dao;
 
 import org.springframework.data.repository.CrudRepository;
+//import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.jakyjty.model.user.User;
 
-public interface IUserDao extends CrudRepository<User, Integer> {
+public interface IUserDao extends CrudRepository<User, Long> {
 
 }
