+++ Jakyjty/src/main/java/com/jakyjty/user/managemService/IUserManagementService.java
@@ -6,6 +6,6 @@ public interface IUserManagementService {
 	
 	
 	public String registerUser(User user);
-	public User fetchUserById(int id);
+	public User fetchUserById(Long id);
 
 }
